+++ src/goose/api/scheduled_jobs.clj
@@ -1,5 +1,5 @@
 (ns goose.api.scheduled-jobs
-  "API to manage scheduled jobs.\\
+  "API to manage scheduled jobs.
   - [API wiki](https://github.com/nilenso/goose/wiki/API)"
   (:require
     [goose.broker :as b]))