+++ src/goose/specs.clj
@@ -109,27 +109,27 @@
 ;;; ============== Retry Opts ==============
 (s/def ::max-retries nat-int?)
 (s/def ::retry-queue (s/nilable ::queue))
-(s/def ::skip-dead-queue boolean?)
 (s/def ::retry-delay-sec-fn-sym
   (s/and ::fn-sym #(some #{1} (u/arities %))))
 (s/def ::error-handler-fn-sym
   (s/and ::fn-sym #(some #{3} (u/arities %))))
 (s/def ::death-handler-fn-sym
   (s/and ::fn-sym #(some #{3} (u/arities %))))
+(s/def ::skip-dead-queue boolean?)
 (s/def ::retry-opts
   (s/and
     (s/map-of #{:max-retries
                 :retry-delay-sec-fn-sym
-                :skip-dead-queue
                 :retry-queue
                 :error-handler-fn-sym
-                :death-handler-fn-sym}
+                :death-handler-fn-sym
+                :skip-dead-queue}
               any?)
     (s/keys :req-un [::max-retries
                      ::retry-delay-sec-fn-sym
-                     ::skip-dead-queue
                      ::error-handler-fn-sym
-                     ::death-handler-fn-sym]
+                     ::death-handler-fn-sym
+                     ::skip-dead-queue]
             :opt-un [::retry-queue])))
 
 ;;; ============== StatsD Metrics ==============
@@ -230,7 +230,7 @@
 (s/fdef goose.brokers.redis.broker/new-producer
         :args (s/cat :redis ::redis-conn-opts)
         :ret ::broker)
-(defn assert-redis-producer [conn-opts]
+(defn ^:no-doc assert-redis-producer [conn-opts]
   (assert-specs 'goose.brokers.redis.broker/new-producer ::redis-conn-opts conn-opts))
 
 (s/fdef goose.brokers.redis.broker/new-consumer
@@ -238,7 +238,7 @@
                      :two (s/cat :redis ::redis-conn-opts
                                  :scheduler-polling-interval-sec ::redis-scheduler-polling-interval-sec))
         :ret ::broker)
-(defn assert-redis-consumer
+(defn ^:no-doc assert-redis-consumer
   [conn-opts scheduler-polling-interval-sec]
   (assert-specs 'goose.brokers.redis.broker/new-consumer ::redis-conn-opts conn-opts)
   (assert-specs 'goose.brokers.redis.broker/new-consumer
@@ -250,7 +250,7 @@
                      :two (s/cat :opts ::rmq
                                  :channels pos-int?))
         :ret ::broker)
-(defn assert-rmq-producer
+(defn ^:no-doc assert-rmq-producer
   [opts channels]
   (assert-specs 'goose.brokers.rmq.broker/new-producer ::rmq opts)
   (assert-specs 'goose.brokers.rmq.broker/new-producer pos-int? channels))
@@ -258,16 +258,16 @@
 (s/fdef goose.brokers.rmq.broker/new-consumer
         :args (s/cat :opts ::rmq)
         :ret ::broker)
-(defn assert-rmq-consumer [opts]
+(defn ^:no-doc assert-rmq-consumer [opts]
   (assert-specs 'goose.brokers.rmq.broker/new-consumer ::rmq opts))
 
 (s/fdef goose.metrics.statsd/new
         :args (s/cat :opts ::statsd-opts)
         :ret ::metrics-plugin)
-(defn assert-statsd [opts]
+(defn ^:no-doc assert-statsd [opts]
   (assert-specs 'goose.metrics.statsd/new ::statsd-opts opts))
 
 (s/fdef goose.worker/start
         :args (s/cat :opts ::worker-opts))
-(defn assert-worker [opts]
+(defn ^:no-doc assert-worker [opts]
   (assert-specs 'goose.worker/start ::worker-opts opts))